Add option to provide channel name directly in Jira webhook
Instead of mapping the project key to a channel, use provided channel name as configured in Jira webhook. This requires to configure a Jira webhook per channel but removes the need to manage the channel mapping configuration on the bridge side.  Also, it allows having the same project post events into multiple channels.